+++ test/transform/resource/before/SuperBuilderNestedGenericTypes.java
@@ -1,9 +1,9 @@
 public class SuperBuilderNestedGenericTypes {
-	@lombok.experimental.SuperBuilder
+	@lombok.SuperBuilder
 	public static abstract class Generic<T extends Generic<?>> {
 	}
 	
-	@lombok.experimental.SuperBuilder
+	@lombok.SuperBuilder
 	public static abstract class NestedGeneric<T extends OtherGeneric<?>> extends Generic<NestedGeneric<? extends OtherGeneric<?>>> {
 	}