+++ test/transform/resource/before/SuperBuilderWithCustomBuilderMethod.java
@@ -2,13 +2,13 @@
 import java.util.List;
 
 public class SuperBuilderWithCustomBuilderMethod {
-	@lombok.experimental.SuperBuilder
+	@lombok.SuperBuilder
 	public static class Parent<A> {
 		A field1;
 		@lombok.Singular List<String> items;
 	}
 	
-	@lombok.experimental.SuperBuilder
+	@lombok.SuperBuilder
 	public static class Child<A> extends Parent<A> {
 		double field3;
 		public static <A> ChildBuilder<A, ?, ?> builder() {