+++ test/transform/resource/before/SuperBuilderWithGenerics3.java
@@ -1,12 +1,12 @@
 import java.util.List;
 
 public class SuperBuilderWithGenerics3 {
-	@lombok.experimental.SuperBuilder
+	@lombok.SuperBuilder
 	public static class Parent<A> {
 		private final String str;
 	}
 	
-	@lombok.experimental.SuperBuilder
+	@lombok.SuperBuilder
 	public static class Child extends Parent<Child.SomeInnerStaticClass> {
 		public static class SomeInnerStaticClass { }
 		double field3;